Allow user to provide the fclose() function to close a FILE object
This is probably unlikely to be really useful, but it simply replaces
the boolean to choose whether to close or not the stream, and allow
for example to pass an hypothetical g_fclose() function to close a
g_fopen()-ed object or so. In one word, it's only more extensible.

This breaks both API and ABI.