Fix EOF flag setting in mio_gets() when \n is at the end of file
When the input file looks like

abcd\nEOF

and mio_gets() is called, it returns "abcd\n" and sets the EOF flag.
However, fgets() doesn't set EOF in this case - it only sets it in the
subsequent call. Fix this.

The unit test input file had to be modified to always put \n at the end
of the file in order to test this issue.